add redundant cast

+++ tools/jxlinfo.c
@@ -340,7 +340,7 @@ int PrintBasicInfo(FILE* file, int verbose) {
         printf("box: type: \"%c%c%c%c\" size: %" PRIu64
                ", contents size: %" PRIu64 "\n",
                type[0], type[1], type[2], type[3], (uint64_t)size,
-               contents_size);
+               (uint64_t)contents_size);
       }
       if (!strncmp(type, "JXL ", 4)) {
         printf("JPEG XL file format container (ISO/IEC 18181-2)\n");
